+++ groupware/CI/application/controllers/member.php
@@ -300,7 +300,9 @@ class Member extends CI_Controller{
 					//이전파일 삭제하고 업로드
 					
 					if($getData['file']){
-						unlink(realpath($config['upload_path']) . '/' . $getData['file']);
+						if( is_file(realpath($config['upload_path']) . '/' . $getData['file']) ){
+							unlink(realpath($config['upload_path']) . '/' . $getData['file']);
+						}
 					}
 					$upload_data = $this->upload->data();
 					$file = $upload_data['file_name'];
@@ -348,7 +350,9 @@ class Member extends CI_Controller{
 			
 			foreach( $list as $lt ){
 				if($lt['file'] != ''){
-					unlink(realpath($config['upload_path']) . '/' . $lt['file']);
+					if( is_file(realpath($config['upload_path']) . '/' . $lt['file']) ){
+						unlink(realpath($config['upload_path']) . '/' . $lt['file']);
+					}
 				}
 			}
 			$this->member_model->set_user_delete($option);
